requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==42.0.7
dataclasses-json==0.6.6
Deprecated==1.2.14
distro==1.9.0
docutils==0.18.1
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
importlib-metadata==4.11.2
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
langchain==0.2.2
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langsmith==0.1.67
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pinecone-client==3.2.2
pkginfo==1.8.2
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyGithub==2.3.0
Pygments==2.11.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.7
pypdf==4.2.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
readme-renderer==34.0
regex==2024.5.15
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
twine==3.8.0
typing-inspect==0.9.0
typing_extensions==4.12.1
ue4cli==0.0.52
urllib3==1.26.8
webencodings==0.5.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.7.0